move GTM integration code to Integrations namespace
